+++ CHANGELOG.md
@@ -22,6 +22,18 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.7.0] - UNRELEASED
+
+### Added
+
+- New `--input-folder` option that scans all `.txt` within a folder for URLs (non recursive). URLs within the same file will be a group;
+  A subfolder within the downloads folder will be created for each group. See: <https://script-ware.gitbook.io/cyberdrop-dl/reference/config/sorting#group-urls>
+- New `--source` option that accepts either a file or a folder
+
+### Changed
+
+- `-i` is now an alias for `--source` instead of `--input-file`
+
 ## [10.6.1] - 2026-08-25
 
 ### Changed